Development (#11)
* adding Kevin Harrington to the authors page

* API for constructor changed

The Constructor takes the serial port number, but in the comments it
notes that Serial1 is the only valid one. THis has been changed to only
access Serial1 and removed the parameter.

* begin refactoring

The competition code is being switched over to a function pointer model.

* Partial new packet parser working

* Switching back to original parser

* fixing the start button detect

* BAsic competition state machine working

* DWF v2 is ready for design review

* Adding DFW pointer to lambdas

The functions that the user defines, now include the passing of the DFW
object from the fundtion call to the user function. THe lets the user
functions to be in isolated files and avoids code clutter. It also
provides an example to the users of how to pass objects around between
sections of code.

* Updating API and packet parsing

The packet parsing is done in line rather than depending on the find
functions.

* Move the update function to a private method

* fix the polarity of button presses

* Add an additional state for waiting for teleop to start

The start button needs to be hit a second time to kick the framework
into teleop mode.

* Adding a status flash

The debug led will now flash when communication is working, and stay
solid on if it fails.

* test code should check for powerup state

* Communications tied to 9600 and single line returns for buttons

The buttons now return the result of the mask check directly. In
addition all of the initialization code loads using 9600 only.

* tab separated prints

* updated comments

* spelling

* Allow all loops to fall through on state change

* formatting and comment removal

* removing internal memory declaration on joystick functions

* Comments and formatting

* Fixed the startup LED state when controller not connected

* formatting

* Bug fixes to clean up warnings in strict mode

* Extended Robot Template with objects

* Updating the robot template code